initial commit as package to install
