+++ README.md
@@ -7,4 +7,3 @@ Run `bash on.sh` so to run the script as a nohup process. Pass args to python
 script as you would running `python ./script.py *args`
 
 Run `bash off.sh` to kill the python script process, and turn off the lights.
-
